requirements.txt
# FastAPI and ASGI server
fastapi==0.104.1
uvicorn==0.23.2
pydantic==2.4.2
python-dotenv==1.0.0

# HTTP client
requests==2.31.0
aiohttp==3.8.6

# Data processing
numpy==1.26.0
pandas==2.1.1
matplotlib==3.8.0
seaborn==0.12.2

# AI and ML
scikit-learn==1.3.1
torch==2.1.0
transformers==4.34.0
sentence-transformers==2.2.2
google-generativeai==0.3.1

# Vector stores
pinecone-client==2.2.4
redis==5.0.1

# Trading APIs
python-binance==1.0.19
ccxt==4.1.13

# Visualization
plotly==5.17.0
dash==2.13.0

# Utilities
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1
python-multipart==0.0.6
tenacity==8.2.3
tqdm==4.66.1

# Testing
pytest==7.4.2
pytest-asyncio==0.21.1
httpx==0.25.0 